# specs/001-youtube-oauth-subscriptions/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: YouTube Health Check - OAuth and Collection API
  version: 1.0.0
  description: |
    Browser-facing contract for feature 001. All endpoints are same-origin. OAuth tokens and
    YouTube subscription IDs never appear in a response. Mutation endpoints require both the
    opaque session cookie and X-CSRF-Token, and reject an Origin different from APP_BASE_URL.
servers:
  - url: http://localhost:3000
paths:
...
  /api/oauth/callback:
    get:
      operationId: completeOAuth
      summary: Validate state, exchange the authorization code, and redirect to the app
      security:
        - cookieAuth: []
      parameters:
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: error
          in: query
          required: false
          schema:
            type: string
      responses:
        '303':
          description: |
            Redirect to /collect on success or to /?oauth=<safe-error-code> on denial/failure.
            The authorization code and state are never copied to the redirect URL.
          headers:
            Location:
              required: true
              schema:
                type: string
        '400':
          $ref: '#/components/responses/InvalidRequest'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sac_session
...
  responses:
    InvalidRequest:
      description: Input or callback validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorView'
...
  schemas:
...
    ErrorView:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        retryAfterMs:
          type: integer
          minimum: 0